requirements.txt
numpy==1.24.4
opencv-python==4.12.0.88
ultralytics==8.3.235
orjson==3.10.15
fastapi==0.124.0
uvicorn[standard]==0.33.0
pillow==10.4.0
easyocr==1.7.2
scikit-image==0.21.0
python-telegram-bot==21.4
httpx==0.27.2
python-multipart==0.0.20

# SAM2 через Ultralytics (автоматическая загрузка весов)
torch==2.1.2
torchvision==0.16.2
scikit-learn==1.3.2
networkx==3.1

